==> 0-preliminaries.ipynb <==
# NASA-hosted Cloud Optimized Geotiffs Demo

Most datasets in NASA's archive are stored as NetCDF, HDF5, or TIFF formats. Occaisionally, NASA Distributed Active Archive Centers (DAACs) will host services such as [OpenDAP](https://earthdata.nasa.gov/collaborate/open-data-services-and-software/api/opendap) to provide server-side subsetting functionality, which becomes increasingly useful as the size and complexity of individual data files grows. These services allow users to circumvent downloading entire files, instead extracting a subregion or subset of observational variables. 

An alternative emerging access pattern is to use data formats that allow client-side subsetting instead of maintaining complicated server applications. [Cloud-Optimized Geotiff (COG)](https://www.cogeo.org) for example, is a format that uses internal tiling in order for client-side software libraries to retrieve only portions of a single file over a network connection. This notebook illustrates how this functionality

### Part 1) Locate data

The entire NASA archive can be programmatically searched via the [Common Metadata Repository (CMR)](https://cmr.earthdata.nasa.gov/search/site/docs/search/api.html). Anyone can search this archive, but to download files you'll need to authenticate using [NASA's Earthdata login](https://wiki.earthdata.nasa.gov/display/EL/How+To+Access+Data+With+cURL+And+Wget) (click on the link for instructions to set up a user name and password, and configure a `~/.netrc` file on your computer)




We'll subset here for a MEaSURES dataset consisting of Greenland Image Mosaics form Sentinel-1A and 1-B Synthetic Aperture Radar Backscatter (https://nsidc.org/data/nsidc-0723). Note this is a large data collection (> 1000 individual files, > 1.5Tb). If you want to explore and analyze this dataset it would be inconvenient to download everything. Nevertheless, NASA happily gives you the option to do so, in the form of a download script. Let's modify that script to just get a list of URLs to COGs without downloading them:

# CMR API Search (Need NASA EarthdataLoging configured in .netrc file locally)
import cmr

short_name = 'NSIDC-0723'
version = '3'
time_start = '2015-01-01T00:00:00Z'
time_end = '2020-10-05T15:43:33Z'
#bounding_box = '-54.85,69.31,-52.18,70.26'
bounding_box = None
polygon = None
#filename_filter = '*gamma*'
filename_filter = None

urls = cmr.get_urls(short_name, version, time_start, time_end, bounding_box, polygon, filename_filter)

print(f'retrieved {len(urls)} URLs')

# Note these URLs include JPG thumbnails, full resolution TIF, metadata in the form of XML
urls[:5]

# Get URLs for all JPG thumbnails
jpgs = [url for url in urls if url.endswith('jpg')]
jpgs[:5]

# Filter out COG urls only 
cogs = [url for url in urls if url.endswith('tif')]
print(len(cogs))
cogs[:5]

# filter 'gamma', 'sigma', and 'image'
gammas = [cog for cog in cogs if 'gamma' in cog]
images = [cog for cog in cogs if 'image' in cog]
sigmas = [cog for cog in cogs if 'sigma' in cog]

# Save these lists to a file for use later:
with open('gamma0.txt', 'w') as f:
    f.write('\n'.join(gammas))

!head gamma0.txt
